Adds held button state

+++ Assets/_Source/Stats/Guns/GunData.cs
@@ -18,7 +18,7 @@ public abstract class GunData : CardStatistics
     public AudioClip BulletEnemyHitSound;
     public bool IsAutomatic;
     public int MagazineSize;
-    public int FireRate;
+    public float FireRate;
     public float Range;
     public float ReloadTime;
     public float BulletDamage;
